Login com token

+++ API/api/components/user/usersController.js
@@ -4,6 +4,7 @@ const { check, validationResult } = require('express-validator/check');
 
 var usersService = require('./usersService')
 var bcrypt = require('bcryptjs');
+var jwt = require('jsonwebtoken');
 
 //criação de um novo utilizador
 exports.createUser = async (req, res) => {
@@ -43,12 +44,12 @@ exports.createUser = async (req, res) => {
     }
     //caso contrário, cria o utilizador
     else {
-        var hash = bcrypt.hashSync(req.body.hashPassword,8);
+        var hash = bcrypt.hashSync(req.body.hashPassword, 8);
         var newUser = {
-            email:req.body.email,
-            username:req.body.username,
-            nome:req.body.nome,
-            hashPassword:hash
+            email: req.body.email,
+            username: req.body.username,
+            nome: req.body.nome,
+            hashPassword: hash
         }
         var createUser;
         //criação de um novo user de acordo com os parâmetros recebidos
@@ -64,19 +65,52 @@ exports.getUser = async (req, res) => {
     let serverResponse = { status: "Não existe na base de dados", response: {} }
     var getUser;
     const userName = req.params.username;
-    await usersService.getUser(userName).then(user => getUser = user).catch(err => console.log(err));
-    if (getUser != null) {
-        serverResponse = { status: "Utilizador existe na Base de Dados", response: getUser }
+    //token recebido
+    var token = req.headers['x-access-token'];
+    //se não existir token
+    if (!token) {
+        return res.status(401).send({ auth: false, message: 'No token provided.' });
+    }
+    //se existir 
+    try {
+        //validar
+        jwt.verify(token, 'secret');
+        await usersService.getUser(userName).then(user => getUser = user).catch(err => console.log(err));
+        if (getUser != null) {
+            serverResponse = { status: "Utilizador está na base de dados", response: getUser }
+        }
+        res.send(serverResponse);
+    } catch (err) {
+        res.status(500).send({ auth: false, message: 'Failed to authenticate token.' });
     }
 
-    res.send(serverResponse);
 }
 
 //request a todos os utilizadores
 exports.getAllUsers = async (req, res) => {
+    let serverResponse = { status: "Não existe utilizadores", response: {} }
+    //variável que recebe a query da base de dados
     var allUsers;
-    await usersService.getAllUsers().then(user => allUsers = user).catch(err => console.log(err));
-    res.send(allUsers);
+    //token
+    var token = req.headers['x-access-token'];
+    //se o token não existir
+    if (!token) {
+        return res.status(401).send({ auth: false, message: 'No token provided.' });
+    }
+    //caso exista
+    try {
+        //validar
+        jwt.verify(token, 'secret');
+        await usersService.getAllUsers().then(user => allUsers = user).catch(err => console.log(err));
+        if (allUsers != null) {
+            serverResponse = { status: "Existem utilizadores na Base de Dados", response: allUsers }
+            res.send(serverResponse);
+        }
+
+    } catch (err) {
+        res.status(500).send({ auth: false, message: 'Failed to authenticate token.' });
+
+    }
 }
 
 //editar um utilizador
@@ -85,13 +119,34 @@ exports.editUser = async (req, res) => {
     let serverResponse = { status: "Not Updated | Utilizador não está na base de dados", response: {} }
     //username inserido no URL
     const userName = req.params.username;
-    await usersService.editUser(req.body, userName).then(user => userUpdate = user).catch(err => console.log(err));
-
-    if (userUpdate != 0) {
-        serverResponse = { status: "Updated", response: userUpdate }
+    //hash da nova password inserida
+    var hash = bcrypt.hashSync(req.body.hashPassword, 8);
+    //user a atualizar
+    var updateUser = {
+        email: req.body.email,
+        username: req.body.username,
+        nome: req.body.nome,
+        hashPassword: hash
     }
-
-    res.send(serverResponse);
+    //token
+    var token = req.headers['x-access-token'];
+    //se o token não existir
+    if (!token) {
+        return res.status(401).send({ auth: false, message: 'No token provided.' });
+    }
+    //se existir
+    try {
+        //verificar
+        jwt.verify(token, 'secret');
+        //update à base de dados
+        await usersService.editUser(updateUser, userName).then(user => userUpdate = user).catch(err => console.log(err));
+        if (userUpdate != 0) {
+            serverResponse = { status: "Updated", response: userUpdate }
+        }
+        res.send(serverResponse);
+    } catch (err) {
+        res.status(500).send({ auth: false, message: 'Failed to authenticate token.' });
+    }
 }
 
 //eliminar utilizador
@@ -100,9 +155,47 @@ exports.deleteUser = async (req, res) => {
     var userDelete;
     //username inserido no URL
     const userName = req.params.username;
-    await usersService.deleteUser(userName).then(user => userDelete = user).catch(err => console.log(err));
-    if (userDelete != 0) {
-        serverResponse = { status: "Deleted", response: userDelete }
+    //token
+    var token = req.headers['x-access-token'];
+    //se o token não existir
+    if (!token) {
+        return res.status(401).send({ auth: false, message: 'No token provided.' });
+    }
+    //se existir
+    try {
+        //validar
+        jwt.verify(token, 'secret');
+        //apagar utilizador
+        await usersService.deleteUser(userName).then(user => userDelete = user).catch(err => console.log(err));
+        if (userDelete != 0) {
+            serverResponse = { status: "Deleted", response: userDelete }
+        }
+        res.send(serverResponse);
+    } catch (err) {
+        res.status(500).send({ auth: false, message: 'Failed to authenticate token.' });
+    }
+}
+
+exports.login = async (req, res) => {
+    let serverResponse = { status: "Não Autenticado", response: {}, token: {} }
+    var existsUserName;
+    var username = req.body.username;
+    //verficar se existe algum utilizador na base de dados com o  username inserido
+    await usersService.getUser(username).then(user => existsUserName = user).catch(err => console.log(err));
+    //se não existir esse utilizador ou a password estiver errada
+    if (existsUserName == null || !bcrypt.compareSync(req.body.password, existsUserName.hashPassword)) {
+        serverResponse = { status: "Username ou password errados", response: {}, token: {} };
+        return res.send(serverResponse);
+    }
+    //se existir o utilizador e a password bater certo
+    else {
+        // create a token
+        var token = jwt.sign({ id: username }, 'secret', {
+            expiresIn: 120 // expires in 2 minutos ***PARA TESTES****
+
+        });
+        serverResponse = { status: "Autenticado", response: existsUserName, token: token }
+
     }
     res.send(serverResponse);
 }
